Preivew paste upload
