+++ assets/src/edit-story/components/canvas/canvasUploadDropTarget.js
@@ -13,12 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/**
- * External dependencies
- */
-import { useCallback } from 'react';
-
 /**
  * WordPress dependencies
  */
@@ -33,78 +27,17 @@ import {
   UploadDropTargetMessage,
   UploadDropTargetOverlay,
 } from '../uploadDropTarget';
-import { useMedia } from '../../app/media';
-import { useStory } from '../../app/story';
-import useInsertElement from '../canvas/useInsertElement';
-import objectPick from '../../utils/objectPick';
 
 import { Layer as CanvasLayer, PageArea } from './layout';
+import useUploadWithPreview from './useUploadWithPreview';
 
 const MESSAGE_ID = 'edit-story-canvas-upload-message';
 
 function CanvasUploadDropTarget({ children }) {
-  const {
-    actions: { uploadMedia, uploadVideoPoster },
-  } = useMedia();
-  const insertElement = useInsertElement();
-  const {
-    actions: { updateElementById, deleteElementById },
-  } = useStory();
+  const uploadWithPreview = useUploadWithPreview();
 
-  const onLocalFile = useCallback(
-    ({ resource }) => {
-      const element = insertElement(resource.type, { resource });
-      return element;
-    },
-    [insertElement]
-  );
-
-  const onUploadedFile = useCallback(
-    async ({ resource, element }) => {
-      const keysToUpdate = objectPick(resource, [
-        'src',
-        'width',
-        'height',
-        'length',
-        'lengthFormatted',
-        'id',
-      ]);
-      const updatedResource = {
-        ...element.resource,
-        ...keysToUpdate,
-      };
-      updateElementById({
-        elementId: element.id,
-        properties: {
-          resource: updatedResource,
-        },
-      });
-      if (resource.type === 'video') {
-        await uploadVideoPoster(resource.id, resource.src);
-      }
-    },
-    [updateElementById, uploadVideoPoster]
-  );
-
-  const onUploadFailure = useCallback(
-    ({ element }) => {
-      deleteElementById({ elementId: element.id });
-    },
-    [deleteElementById]
-  );
-
-  const onDropHandler = useCallback(
-    (files) => {
-      uploadMedia(files, {
-        onLocalFile,
-        onUploadedFile,
-        onUploadFailure,
-      });
-    },
-    [uploadMedia, onLocalFile, onUploadedFile, onUploadFailure]
-  );
   return (
-    <UploadDropTarget onDrop={onDropHandler} labelledBy={MESSAGE_ID}>
+    <UploadDropTarget onDrop={uploadWithPreview} labelledBy={MESSAGE_ID}>
       {children}
       <UploadDropTargetOverlay>
         <CanvasLayer>
